src/inkywhat_dsp/requirements.txt
rich==13.6.0
more-itertools==10.1.0
Pillow==10.0.1
fastapi==0.103.2
pydantic==2.4.2
inky[rpi]==1.5.0
torchvision==0.16.0
